Refactor PDF retrieval methods in box module
Refactored two methods in the box module for getting PDFs. Separated the functionality to not only get the PDFs from a folder, but also to get a flattened version of them, which includes their respective PNG files. This update enhances the module functionalities and optimizes the code for better maintainability.